Rename transactions to payments
unifiying the name
