+++ src/extension/background-script/actions/payments/all.js
@@ -3,14 +3,11 @@ import db from "../../db";
 const all = async (message, sender) => {
   // TODO load transactions from the node and merge it
   // with the payments data stored locally
-  let transactions = await db.payments
-    .toCollection()
-    .reverse()
-    .sortBy("createdAt");
+  let payments = await db.payments.toCollection().reverse().sortBy("createdAt");
 
   return {
     data: {
-      transactions,
+      payments,
     },
   };
 };
