+++ src/extension/background-script/actions/payments/index.js
